Create separate host options for browser and native envs
